various updates, inc language icons and replay buttons
