+++ packages/react/src/IDE/IDE.module.css
@@ -192,6 +192,10 @@
   border-color: var(--brand-color-border-default);
 }
 
+.IDE__Chat-input-area svg {
+  fill: var(--brand-color-text-muted);
+}
+
 .IDE__Chat-message-content {
   margin-top: var(--base-size-16);
 }
@@ -268,6 +272,8 @@
   color: var(--brand-color-text-muted);
   cursor: pointer;
   border: none;
+  transition: opacity var(--brand-animation-duration-default) var(--brand-animation-easing-default);
+  opacity: 1;
 }
 
 .IDE--default .IDE__Editor-tab {
@@ -294,11 +300,24 @@
   color: var(--brand-color-text-default);
 }
 
+.IDE__Editor-tab[disabled] {
+  cursor: not-allowed;
+  color: var(--brand-color-text-subtle) !important;
+  opacity: 0.7;
+}
+
 .IDE__Editor-tab-close-icon {
   width: 12px;
   height: 12px;
   margin-inline-start: var(--base-size-4);
   fill: var(--brand-color-text-muted);
+}
+
+img.IDE__Editor-tab-icon {
+  width: 16px;
+  height: 16px;
+  -webkit-filter: grayscale(10%);
+  filter: grayscale(10%);
 }
 
 .IDE__Editor-tab:hover .IDE__Editor-tab-close-icon {
@@ -425,4 +444,9 @@ pre.IDE__Chat-copilot-indicator {
   align-self: center;
   position: relative;
   z-index: 1;
+}
+
+.IDE__Editor-replay[disabled] {
+  opacity: 0.5;
+  cursor: not-allowed;
 }
