+++ packages/react/src/IDE/IDE.stories.module.css
@@ -1,8 +1,9 @@
-/* .vfx {
+.perspective {
+  max-width: 1024px;
   -webkit-mask-image: linear-gradient(to right, var(--brand-color-canvas-default) 10%, transparent 80%);
   mask-image: linear-gradient(to right, var(--brand-color-canvas-default) 10%, transparent 80%);
-  transform: perspective(220px) rotate(2deg) rotateY(-4deg);
-} */
+  transform: perspective(320px) rotateX(0deg) rotateY(-4deg);
+}
 
 .riverVisual {
   background-repeat: no-repeat;
