+++ packages/react/src/IDE/IDE.stories.module.css.d.ts
@@ -1,4 +1,5 @@
 declare const styles: {
+  readonly "perspective": string;
   readonly "riverVisual": string;
 };
 export = styles;
